requirements.txt
backcall==0.1.0
click==6.7
decorator==4.3.0
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.6
Flask-WTF==0.14.2
ipython==6.4.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
MarkupSafe==1.0
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
PyYAML==3.13
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.2.1
